[build resilience fix] Install DB packages via versioned names
MariaDB and MySQL package switched to the versioned package layout in Fedora and RHEL 10.
The packages can be available in various states, but in all case where a specific version is wanted, specific version must be requested.

Thus container wanting MariaDB 10.11 specifically, must request to install 'mariadb10.11-server' specifically,
instead of just 'mariadb-server' which will give you any version that is selected to be the distribution default the current moment.

That creates an issue with 'rpm -V', as it does not accepts virtual provides as package names to process, failing.
Because of that, we have to move the DB server out of the ${INSTALL_PKGS} variable,
and set up second 'rpm -V' execution that process all of the DB RPMs installed.

That is actually an enhancement, as before it would only check the '*-server' that is specifically listed in the DNF command,
even though more sup-packages of the DB are installed and they are worth verifying too.